rename: 패키지명 수정
- GameApplication -> Application 이름 변경
- RacingGame -> CarRacingController 이름 변경
- generator 패키지는 model 패키지로 이동
