+++ src/main/java/org/duckstudy/Application.java
@@ -3,16 +3,16 @@ package org.duckstudy;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import org.duckstudy.controller.RacingGame;
+import org.duckstudy.controller.CarRacingController;
 import org.duckstudy.controller.Validator;
-import org.duckstudy.generator.Generator;
-import org.duckstudy.generator.RandomValueGenerator;
+import org.duckstudy.model.generator.Generator;
+import org.duckstudy.model.generator.RandomValueGenerator;
 import org.duckstudy.view.InputView;
 import org.duckstudy.view.OutputView;
 
-public class GameApplication {
+public class Application {
 
-    private GameApplication() {
+    private Application() {
     }
 
     public static void main(String[] args) throws IOException {
@@ -22,7 +22,7 @@ public class GameApplication {
         Validator validator = new Validator(outputView);
         Generator generator = new RandomValueGenerator();
 
-        RacingGame racingGame = new RacingGame(inputView, validator, outputView, generator);
-        racingGame.run();
+        CarRacingController carRacingController = new CarRacingController(inputView, validator, outputView, generator);
+        carRacingController.run();
     }
 }
