+++ src/main/java/org/duckstudy/model/generator/Generator.java
@@ -1,4 +1,4 @@
-package org.duckstudy.generator;
+package org.duckstudy.model.generator;
 
 public interface Generator {
     int generateValue();
